Changed how messages was read in canbus

+++ CanBusApp/src/main/java/Main/CAN_CODES.java
@@ -9,8 +9,8 @@ public class CAN_CODES {
     /*-------------Modes-------------*/
     public static final int MODE_SCD = 0x01; // Show current data
     /*-------------PID---------------*/
-    public static int ENGINE_COOL_TEMP; // Engine Coolant temperature.
-    public static int ENGINE_RPM; // Engine RPM
+    public static byte ENGINE_COOL_TEMP; // Engine Coolant temperature.
+    public static byte ENGINE_RPM; // Engine RPM
     /*-------------Type of message-------------*/
     public static final int REQUEST_MSG = 0x7DF;
     public static final int[] RESPONSE_MSG = {0x7E8, 0x7E9, 0x7EA, 0x7EB, 0x7EC, 0x7ED, 0x7EE, 0x7EF};
@@ -48,7 +48,7 @@ public class CAN_CODES {
      * @param pid
      */
     public void assignVariables(JSONObject pid){
-        ENGINE_RPM = Integer.decode(((JSONObject) pid.get("rpm")).get("code").toString());
-        ENGINE_COOL_TEMP = Integer.decode(((JSONObject) pid.get("enginecooltemp")).get("code").toString());
+        ENGINE_RPM = Byte.decode(((JSONObject) pid.get("rpm")).get("code").toString());
+        ENGINE_COOL_TEMP = Byte.decode(((JSONObject) pid.get("enginecooltemp")).get("code").toString());
     }
 }
